requirements.txt
<<<<<<< HEAD
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
imutils==0.5.4
PyYAML==6.0.1
Pillow==10.1.0
matplotlib==3.7.1
pandas==2.0.3
sqlalchemy==2.0.23
python-dotenv==1.0.0
=======
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv
opencv-python
numpy
PyYAML
Pillow
scipy
imutils
scikit-image
>>>>>>> 99f3ab2... : Initial commit: Water level measurement system
